Moving from the GS to the Urban model of MSC togetehr with the related modifications in the electron mamager.
